requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.6
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
executing==2.1.0
inflection==0.5.1
ipython==8.30.0
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
setuptools==75.6.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
wcwidth==0.2.13
